Blog_Backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
